+++ package.json
@@ -27,10 +27,8 @@
     "babel-traverse": "^6.26.0",
     "compose-function": "^3.0.3",
     "dependency-cruiser": "^12.10.0",
-    "effector": "^22.5.2",
     "effector-logger": "^0.13.7",
     "effector-react": "^22.5.0",
-    "eslint": "^8.35.0",
     "eslint-config-prettier": "^8.7.0",
     "eslint-plugin-effector": "^0.10.3",
     "eslint-plugin-jest": "^27.2.1",
@@ -47,5 +45,9 @@
     "vite-plugin-babel": "^1.1.3"
   },
   "main": "index.js",
-  "license": "MIT"
+  "license": "MIT",
+  "devDependencies": {
+    "effector": "^22.5.2",
+    "eslint": "^8.35.0"
+  }
 }
