+++ src/pages/profile-page/ui/index.ts
@@ -1,0 +1,1 @@
+export * from "./profile-page";
