+++ src/pages/routes/ui/index.ts
@@ -1,0 +1,1 @@
+export { Routing } from "./Routing";
