+++ src/shared/config/auth0Config.ts
@@ -1,4 +1,4 @@
-export const auth = {
+export const auth0Config = {
   domain: import.meta.env.VITE_AUTH0_DOMAIN,
   clientId: import.meta.env.VITE_AUTH0_CLIENT_ID,
   authorizationParams: {
