+++ src/shared/config/paths.ts
@@ -1,9 +1,4 @@
-import ProfilePage from "pages/profile-page/ui/profile-page";
-
-export const paths = [
-  {
-    path: "/",
-    component: ProfilePage,
-    key: "profile",
-  },
-];
+export const paths = {
+  path: "/",
+  key: "profile",
+};
